har lakt inkapslin på (player, backpack, rarity och items)

+++ Vinterprojekt-Prog2/Enemy.cs
@@ -1,19 +1,21 @@
 public class Enemy
 {
-    private float maxHP;
-    private float hp;
-    private float damage;
-    private float armor;
+    private double maxHP;
+    private double hp;
+    private double damage;
+    private double armor;
     private int goldDrop;
     private int xpDrop;
 
+
+
     public void Attack()
     {
-        
+
     }
 
     public void Defend()
     {
-        
+
     }
 }
